+++ src/appleseed/renderer/api/postprocessing.h
@@ -37,6 +37,6 @@
 #include "renderer/modeling/postprocessingstage/postprocessingstagetraits.h"
 #include "renderer/modeling/postprocessingstage/renderstamppostprocessingstage.h"
 #include "renderer/modeling/postprocessingstage/vignettepostprocessingstage.h"
-#include "renderer/modeling/postprocessingeffect/ipostprocessingeffect.h"
-#include "renderer/modeling/postprocessingeffect/postprocessingeffectjob.h"
-#include "renderer/modeling/postprocessingeffect/vignettepostprocessingeffect.h"
+#include "renderer/modeling/postprocessingstage/effect/ipostprocessingeffect.h"
+#include "renderer/modeling/postprocessingstage/effect/postprocessingeffectjob.h"
+#include "renderer/modeling/postprocessingstage/effect/vignettepostprocessingeffect.h"
