+++ src/appleseed/renderer/modeling/postprocessingstage/effect/ipostprocessingeffect.h
