+++ src/appleseed/renderer/modeling/postprocessingstage/effect/postprocessingeffectjob.cpp
