+++ src/appleseed/renderer/modeling/postprocessingstage/effect/vignettepostprocessingeffect.cpp
